Core: Processed feedback from PR
