--- src/enums/Granularity.ts
@@ -1,10 +1,0 @@
-/**
- * Granularity types for analytics.
- *
- * @public
- */
-export enum EGranularity {
-	DAILY = 'Daily',
-	WEEKLY = 'Weekly',
-	MONTHLY = 'Monthly',
-}
