+++ src/index.ts
@@ -4,8 +4,7 @@ export * from './Request';
 // ENUMS
 export * from './enums/Media';
 export * from './enums/Notification';
-export * from './enums/Metric';
-export * from './enums/Granularity';
+export * from './enums/Analytic';
 
 // ARG MODELS
 export * from './models/args/NewTweet';
@@ -13,7 +12,6 @@ export * from './models/args/TweetFilter';
 
 // PARAMS MODELS
 export * from './models/params/Variables';
-export * from './models/params/AnalyticOverviewParams';
 
 // RESPONSE TYPES
